refactor: remove old header components and implement new layout structure with blog dropdown and search functionality
