--- src/app/header/layout.tsx
@@ -1,7 +1,0 @@
-import { ReactElement } from "react";
-
-const NavigationMenuDemoLayout = ({ children }: { children: ReactElement }) => {
-  return children
-};
-
-export default NavigationMenuDemoLayout;
