--- src/app/header/utils/menuItems.ts
@@ -1,10 +1,0 @@
-// utils/menuItems.js
-export const menuItems = [
-  { title: "Giới thiệu", href: "/gioi-thieu", color: "text-purple-500", icon: "📘" },
-  { title: "Dịch vụ", href: "/dich-vu", color: "text-blue-500", icon: "💼" },
-  { title: "Dự án", href: "/du-an", color: "text-orange-500", icon: "🏗️" },
-  { title: "Khách hàng", href: "/khach-hang", color: "text-green-500", icon: "👥" },
-  { title: "Blog", href: "/blog", color: "text-red-500", icon: "📝" },
-  { title: "Hoạt động", href: "/hoat-dong", color: "text-yellow-500", icon: "🎉" },
-  { title: "Hợp tác với MONA", href: "/hop-tac", color: "text-pink-500", icon: "🤝" },
-];
