+++ src/app/page.tsx
@@ -1,9 +1,10 @@
 import Image from "next/image";
-import { NavigationMenuDemo } from "./header/page";
 
 export default function Home() {
   return (
-    <NavigationMenuDemo/>
+<div className="">
+  
+</div>
   );
 }
